requirements.txt
asgiref==3.5.0
attrs==21.4.0
black==22.3.0
click==8.1.2
Django==4.0.4
django-extensions==3.1.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-spectacular==0.22.0
inflection==0.5.1
jsonschema==4.4.0
model-bakery==1.5.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
PyJWT==2.3.0
pyrsistent==0.18.1
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
sqlparse==0.4.2
tomli==2.0.1
uritemplate==4.1.1
